Refactor financial structs and connectivity
